fix: SQLite long-to-int runtime errors + xhttp_extra column
- Add explicit (long) and (int) casts in all Dapper mapping methods
- SQLite INTEGER returns long but model properties use int
- Add xhttp_extra column to servers table schema + migration
- Include xhttp_extra in SELECT/INSERT/UPDATE queries

+++ src/AIXray.Storage/GroupRepository.cs
@@ -96,13 +96,13 @@ public class GroupRepository : IGroupRepository
 
     private static Group MapRow(dynamic r) => new()
     {
-        Id = r.id,
+        Id = (long)r.id,
         Name = r.name ?? "",
         SubscriptionUrl = r.subscription_url,
-        AutoUpdate = r.auto_update != 0,
-        UpdateIntervalMinutes = r.update_interval_min,
+        AutoUpdate = (long)r.auto_update != 0,
+        UpdateIntervalMinutes = (int)(long)r.update_interval_min,
         LastUpdate = ParseDateTime(r.last_update),
-        IsDefault = r.is_default != 0,
+        IsDefault = (long)r.is_default != 0,
     };
 
     private static DateTime? ParseDateTime(object? value)
